+++ CHANGELOG.md
@@ -9,7 +9,13 @@ nav_order: 1
 <!-- Always keep the following header in place: -->
 <!--## [MAJOR.MINOR.PATCH] - YYYY-MM-DD -->
 
+## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
+
+- Change `aiven_opensearch` field `opensearch_user_config.opensearch.auth_failure_listeners.ip_rate_limiting`: deprecate:
+  This property is deprecated
+
 ## [4.34.1] - 2025-02-07
+
 - Fix `organization_user_group` import
 
 ## [4.34.0] - 2025-01-29